+++ adapters/kitlogadapter/logger_test.go
@@ -8,7 +8,7 @@ import (
 
 	"github.com/go-kit/kit/log"
 	"github.com/goph/logur"
-	logtesting "github.com/goph/logur/testing"
+	"github.com/goph/logur/logtesting"
 )
 
 func newTestSuite() *logtesting.LoggerTestSuite {