+++ example/logger_test.go
@@ -5,7 +5,7 @@ import (
 	"testing"
 
 	"github.com/goph/logur"
-	logtesting "github.com/goph/logur/testing"
+	"github.com/goph/logur/logtesting"
 )
 
 func TestLoggerAdapter_Levels(t *testing.T) {