+++ logger_context_test.go
@@ -4,7 +4,7 @@ import (
 	"testing"
 
 	. "github.com/goph/logur"
-	logtesting "github.com/goph/logur/testing"
+	"github.com/goph/logur/logtesting"
 )
 
 func newContextualLoggerTestSuite() *logtesting.LoggerTestSuite {