+++ logger_testing_test.go
@@ -5,7 +5,7 @@ import (
 	"testing"
 
 	. "github.com/goph/logur"
-	logtesting "github.com/goph/logur/testing"
+	"github.com/goph/logur/logtesting"
 )
 
 func TestAssertLogEventsEqual(t *testing.T) {